consolidated getinvolved to one module. table now appears for mobile, still requires styling edits
